refractor: fileConversion functions -> fileConversionUtil.ts, preview image functions -> previewImageUtil.ts
