+++ frontend/src/App.tsx
@@ -15,7 +15,6 @@ import { HistoryPage } from "./pages/History/HistoryPage";
 import { HomePage } from "./pages/Homepage/HomePage"
 import { Footer } from "./components/Footer/Footer";
 import ErrorPage from "./pages/ErrorPage/ErrorPage";
-import { generatePreviewImages } from "./util/helper";
 import { useBGLoadingMessage } from "./components/Contexts/LoadingBGMessageContext";
 import { AlertPopup } from "./components/MenuItems/Popups/AlertPopup";
 import { useCustomAlert } from "./components/Hooks/useCustomAlert";
@@ -23,6 +22,7 @@ import { ClerkAuthGuard } from "./components/Feature/Auth/ClerkAuthGaurd";
 import { ClerkProvider, SignIn, useAuth, useUser } from "@clerk/clerk-react";
 import SignInPage from "./pages/LoginPortal/sign-in";
 import { getToken } from "./util/apiUtil";
+import { generatePreviewImages } from "./util/previewImageUtil";
 
 export default function App() {
   const [tracks, setTracks] = useState<TrackResponse[] | null>(null);
