+++ frontend/src/util/helper.ts
@@ -8,7 +8,6 @@ import { BootcampResponse, pdfGenerationResponse, SaltData, Size, Student, stude
 import { useLoadingMessage } from "../components/Contexts/LoadingMessageContext";
 import JSZip from 'jszip';
 import { saveAs } from 'file-saver';
-import * as pdfjsLib from 'pdfjs-dist';
 import { getFontsData } from "./fontsUtil";
 import { api } from "./apiUtil";
 
@@ -87,77 +86,6 @@ export const getPlugins = () => {
     Signature: plugins.signature,
     QR: barcodes.qrcode,
     Image: image,
-  }
-}
-
-// No Longer Used In front-end cause it slows down application. But it is faster than doing it in the backend
-export const convertPDFToImage = async (pdfInput: ArrayBuffer): Promise<Blob | null> => {
-  try {
-    const pdf = await pdfjsLib.getDocument({ data: pdfInput }).promise;
-    
-    const page = await pdf.getPage(1);
-    const viewport = page.getViewport({ scale: 1.5 });
-    const canvas = document.createElement("canvas");
-    canvas.height = viewport.height;
-    canvas.width = viewport.width;
-    
-    const renderContext = {
-      canvasContext: canvas.getContext("2d"),
-      viewport: viewport,
-    };
-    
-    await page.render(renderContext).promise;
-    const dataURL = canvas.toDataURL("image/png");
-    
-    const base64Data = dataURL.replace(/^data:image\/png;base64,/, '');
-    const binaryString = window.atob(base64Data);
-    const len = binaryString.length;
-    const bytes = new Uint8Array(len);
-    
-    for (let i = 0; i < len; i++) {
-      bytes[i] = binaryString.charCodeAt(i);
-    }
-    
-    return new Blob([bytes], { type: 'image/png' });
-  } catch (e) {
-    console.error('Error loading PDF:', e);
-    return null;
-  }
-};
-
-export function convertUint8ArrayToBase64(uint8Array: Uint8Array): string {
-  let binaryString = '';
-  for (let i = 0; i < uint8Array.length; i++) {
-    binaryString += String.fromCharCode(uint8Array[i]);
-  }
-  return btoa(binaryString);
-}
-
-export const generatePreviewImages = async (pdfs: Uint8Array[], students: Student[], setBGLoadingMessage: (message: string) => void, setBootcamps: (response: StudentResponse) => void): Promise<void> => {
-  const pdfConversionRequests: studentImagePreview[] = []; 
-
-  for (let i = 0; i < pdfs.length; i++) {
-    setBGLoadingMessage(`Converting pdfs to blob ${i + 1}/${pdfs.length}`);
-
-    const base64String = convertUint8ArrayToBase64(pdfs[i]);
-
-    await pdfConversionRequests.push({
-      studentGuidId: students[i].guidId,
-      image: base64String
-    });
-  }
-
-  try {
-    for (let i = 0; i < pdfConversionRequests.length; i++) {
-      setBGLoadingMessage(`Converting & Compressing Thumbnails ${i + 1}/${pdfConversionRequests.length}`)
-      const response: StudentResponse = await api.updateStudentPreviewImage(pdfConversionRequests[i])
-      setBootcamps(response);
-    }
-
-    setBGLoadingMessage("Finished!");
-    
-  } catch (error) {
-    setBGLoadingMessage(`Failed to Update PreviewImages!. ${error.message || 'Unknown error'}`)
   }
 }
